Fix reviewer round 2: deduplicate JSON parser, move helper to testing.py
- Rename _parse_agents_from_output to parse_agents_from_mng_output (public)
  and have testing.py's parse_mng_list_json delegate to it, eliminating
  the duplicate JSON parsing logic
- Move make_git_repo from forwarding_server/conftest.py to testing.py
  per project convention (conftest.py is for fixtures, testing.py is for
  explicitly imported helper functions)
